refactor: move errorkind to common
