+++ common/Cargo.toml
@@ -10,11 +10,11 @@ description = "Common library for the shuttle platform (https://www.shuttle.rs/)
 chrono = { version = "0.4.22", features = ["serde"] }
 comfy-table = { version = "6.1.0", optional = true }
 crossterm =  { version = "0.25.0", optional = true }
+http = { version = "0.2.8", optional = true }
 once_cell = "1.13.1"
 rustrict = "0.5.0"
 serde = { version = "1.0.143", features = ["derive"] }
 serde_json = { version = "1.0.85", optional = true }
-http = { version = "0.2.8", optional = true }
 strum = { version = "0.24.1", features = ["derive"] }
 tracing = "0.1.36"
 uuid = { version = "1.1.1", features = ["v4", "serde"] }
