+++ gateway/Cargo.toml
@@ -11,7 +11,6 @@ base64 = "0.13"
 bollard = "0.13"
 chrono = "0.4"
 clap = { version = "4.0.0", features = [ "derive" ] }
-convert_case = "0.5.0"
 fqdn = "0.2.2"
 futures = "0.3.21"
 http = "0.2.8"
